Clustering Characteristics Commit
Clustering is an unsupervised learning algorithm. It is an assignment of a set of observations into subsets (called clusters) so that observations in the same cluster are similar in some manner. Clustering is broadly explained in two types here: Hierarchical Clustering and K-Means Clustering along with their characteristics with 6 case studies